AITEAM-1: Set up tests for sagemaker inference endpoints (#8)
* AITEAM-1: Set up tests for sagemaker inference endpoints

This provides a quasi-template and guide for other engineers to approach
an invocation of their endpoint, and steps to evaluate. The script will
hit an existing endpoint with some AWS provided sample data, then look
at the results. No decisions are made as to the validity of those
results, this would be left as an exercise to the person(s) who are
creating and evaluating their models.

The new job step will run after a dev endpoint has been created, and
before the "gate" for promoted to prod.

* Update kitten_model/README.md

Provide some extra context to the new script, and how it can be used by
other practitioners to do an invoke/eval loop, as appropriate.